fix: move GenerateSizes in separate class
